Fixed hyprland version check
The old expression checking the hyprland version number returned True, if e.g. hyprland v0.35.0 or any v0.3x.0 smaller than v0.37.0 was used. The hyprctl json format of these older versions is not compatible with the current parsing.

Reasoning:
The old "positiv check" (>=v0.37.0) was combined with a logical OR and the "negative check" for versions v0.1.0 and v0.2.0. So if either of this statements is True, the hole expression is true. Therefore e.g. hyprland v0.35.0 also gets True.